+++ test/helpers/setup.ts
@@ -1,4 +1,5 @@
 import Module from 'module';
+import { setOpenUrlForTests } from '../../out/testHooks';
 import { vscodeMock } from './vscodeMock';
 
 const originalRequire = Module.prototype.require;
@@ -9,3 +10,7 @@ Module.prototype.require = function (id: string) {
   }
   return originalRequire.apply(this, arguments as any);
 };
+
+setOpenUrlForTests(() => {
+  // Prevent tests from launching a real browser or editor.
+});